use SDL's hires timer to limit the FPS
it's a little more accurate at least on my system

+++ src/global.h
@@ -96,7 +96,7 @@ typedef struct {
 
 	int frames;
 	int stageuiframes;
-	int lasttime; // frame limiter
+	uint64_t lasttime; // frame limiter
 	int timer;
 	int frameskip;
